+++ packages/styleguide/stories/Organisms/ConnectedForm.stories.mdx
@@ -1,11 +1,15 @@
 import {
+  ConnectedCheckbox,
   ConnectedFormGroup,
+  ConnectedSelect,
   FormWrapper,
   SubmitButton,
 } from '@codecademy/gamut';
 import title from '@codecademy/macros/lib/title.macro';
 import { PropsTable } from '@codecademy/storybook-addon-variance';
 import { Canvas, Meta, Story } from '@storybook/addon-docs/blocks';
+
+import { HEY } from './examples';
 
 <Meta
   title={title}
@@ -22,17 +26,46 @@ import { Canvas, Meta, Story } from '@storybook/addon-docs/blocks';
   }}
 />
 
+<HEY />
+
 <Canvas>
   <Story
     name="FormWrapper"
     args={{ validation: 'onSubmit', resetOnSubmit: true }}
   >
     {(args) => (
-      <FormWrapper onSubmit={() => console.log('hey')}>
-        <ConnectedFormGroup name="this-field" label="hi" size="large">
-          <div /> hey
+      <FormWrapper
+        onSubmit={(values) => console.log(values)}
+        defaultValues={{ 'this-field': true, 'that-field': 'three' }}
+      >
+        <ConnectedFormGroup
+          name="this-field"
+          label="hi"
+          size="large"
+          spacing="tight"
+        >
+          <ConnectedCheckbox
+            label="check me out"
+            name="this-field"
+            htmlFor="this-field"
+            validation={{ required: 'check it...' }}
+          />
         </ConnectedFormGroup>
-        <SubmitButton>submit!!!</SubmitButton>
+        <ConnectedFormGroup name="that-field" label="hey" size="large">
+          <ConnectedSelect
+            label="check me out"
+            name="that-field"
+            htmlFor="that-field"
+            validation={{
+              pattern: {
+                value: /^(?:(?!zero).)*$/,
+                message: 'zero to hero',
+              },
+            }}
+            options={['zero', 'one', 'two', 'three']}
+          />
+        </ConnectedFormGroup>
+        <SubmitButton mt={8}>submit!!!</SubmitButton>
       </FormWrapper>
     )}
   </Story>
