--- packages/styleguide/stories/Organisms/ConnectedForm/About.stories.mdx
@@ -1,76 +1,0 @@
-import {
-  ConnectedCheckbox,
-  ConnectedFormGroup,
-  ConnectedSelect,
-  FormWrapper,
-  SubmitButton,
-} from '@codecademy/gamut';
-import title from '@codecademy/macros/lib/title.macro';
-import { PropsTable } from '@codecademy/storybook-addon-variance';
-import { Canvas, Meta, Story } from '@storybook/addon-docs/blocks';
-
-import { HEY } from './examples';
-
-<Meta
-  title={title}
-  component={FormWrapper}
-  parameters={{
-    subtitle: 'The special sauce to creating modular forms',
-    source: 'gamut',
-    status: 'current',
-    design: {
-      type: 'figma',
-      url:
-        'https://www.figma.com/file/ReGfRNillGABAj5SlITalN/%F0%9F%93%90-Gamut?node-id=1189%3A0',
-    },
-  }}
-/>
-
-<HEY />
-
-<Canvas>
-  <Story
-    name="FormWrapper"
-    args={{ validation: 'onSubmit', resetOnSubmit: true }}
-  >
-    {(args) => (
-      <FormWrapper
-        onSubmit={(values) => console.log(values)}
-        defaultValues={{ 'this-field': true, 'that-field': 'three' }}
-      >
-        <ConnectedFormGroup
-          name="this-field"
-          label="hi"
-          size="large"
-          spacing="tight"
-        >
-          <ConnectedCheckbox
-            label="check me out"
-            name="this-field"
-            htmlFor="this-field"
-            validation={{ required: 'check it...' }}
-          />
-        </ConnectedFormGroup>
-        <ConnectedFormGroup name="that-field" label="hey" size="large">
-          <ConnectedSelect
-            label="check me out"
-            name="that-field"
-            htmlFor="that-field"
-            validation={{
-              pattern: {
-                value: /^(?:(?!zero).)*$/,
-                message: 'zero to hero',
-              },
-            }}
-            options={['zero', 'one', 'two', 'three']}
-          />
-        </ConnectedFormGroup>
-        <SubmitButton mt={8}>submit!!!</SubmitButton>
-      </FormWrapper>
-    )}
-  </Story>
-</Canvas>
-
-<PropsTable of={FormWrapper} />
-
-### Inputs
